🐛 `spatial.cKDTree.count_neighbors`: fix scalar return type

+++ tests/spatial/test__kdtree.pyi
@@ -57,10 +57,10 @@ assert_type(_ctree.query_pairs(1.0, output_type="ndarray"), onp.ArrayND[np.intp]
 
 # cKDTree.count_neighbors
 
-assert_type(_ctree.count_neighbors(_ctree, 1.0), np.intp)
+assert_type(_ctree.count_neighbors(_ctree, 1.0), int)
 assert_type(_ctree.count_neighbors(_ctree, 1.0, 2.0, (_f64_1d, _f64_1d)), np.float64)
 assert_type(_ctree.count_neighbors(_ctree, 1.0, weights=(_f64_1d, _f64_1d)), np.float64)
-assert_type(_ctree.count_neighbors(_ctree, _f64_1d), np.intp | onp.Array1D[np.intp])
+assert_type(_ctree.count_neighbors(_ctree, _f64_1d), int | onp.Array1D[np.intp])
 assert_type(_ctree.count_neighbors(_ctree, _f64_1d, 2.0, (_f64_1d, _f64_1d)), np.float64 | onp.Array1D[np.float64])
 assert_type(_ctree.count_neighbors(_ctree, _f64_1d, weights=(_f64_1d, _f64_1d)), np.float64 | onp.Array1D[np.float64])
 
@@ -134,7 +134,7 @@ assert_type(_tree.query_pairs(1.0, output_type="ndarray"), onp.ArrayND[np.intp])
 
 # KDTree.count_neighbors
 
-assert_type(_tree.count_neighbors(_tree, r=1.0), np.intp)
+assert_type(_tree.count_neighbors(_tree, r=1.0), int)
 assert_type(_tree.count_neighbors(_tree, r=1.0, weights=(_f64_1d, _f64_1d)), np.float64)
 
 # KDTree.sparse_distance_matrix
